修改了我的订单触点问题，添加了再次申请和联系客服功能

+++ app/src/main/java/com/zthx/npj/net/been/LookUserResponseBean.java
@@ -5,7 +5,7 @@ public class LookUserResponseBean extends BaseReponseBean{
         private String nick_name;
         private String head_img;
         private String signature;
-        private long level;
+        private int level;
         private String lat;
         private String lng;
         private String reputation;
@@ -66,11 +66,11 @@ public class LookUserResponseBean extends BaseReponseBean{
             this.signature = signature;
         }
 
-        public long getLevel() {
+        public int getLevel() {
             return level;
         }
 
-        public void setLevel(long level) {
+        public void setLevel(int level) {
             this.level = level;
         }
